Moved user defined hooks to Hooks folder
